Make sure float attributes get properly set
Since the Jsoncpp 'isInt' function returns positively if a double value has a 0 decimal part it could happen that a float attribute was set by using an integer function. I reordered so that floats are set before ints and check on the value type to determine if it's a double rather than using the 'isDouble' function. The 'is*' functions in Jsoncpp really return whether the value can be interpreted as the given type, not whether it is actually stored as the given type which led to this potential issue. This didn't break anything but Arnold outputted warnings.